requirements.txt
certifi==2022.6.15
charset-normalizer==2.1.0
cycler==0.11.0
EasyProcess==1.1
entrypoint2==1.1
fonttools==4.34.4
hurry.filesize==0.9
idna==3.3
kaggle==1.5.12
kiwisolver==1.4.3
matplotlib==3.5.2
numpy==1.23.1
packaging==21.3
patool==1.12
Pillow==9.2.0
pyparsing==3.0.9
python-dateutil==2.8.2
python-slugify==6.1.2
pyunpack==0.3
recordtype==1.3
requests==2.28.1
six==1.16.0
text-unidecode==1.3
tqdm==4.64.0
urllib3==1.26.10
